Move redux logger to module for use outside of E2E.
