+++ tests/e2e/assets/e2e-redux-logger.js
@@ -17,41 +17,9 @@
  */
 
 /**
- * External dependencies
+ * Internal dependencies
  */
-import { applyMiddleware } from 'redux';
+import { setupReduxLogger } from '../../js/redux-debug';
 
-/**
- * Creates a middleware for logging dispatched actions to the console.
- *
- * @return {Function} Middleware function.
- */
-function createLoggerMiddleware() {
-	return ( next ) => ( action ) => {
-		const { type, ..._action } = action;
-
-		// Objects must be stringified to be inspectable from the console during E2E tests.
-		// Not all structures can be stringified so errors must be caught.
-		try {
-			global.console.debug( 'DISPATCH', type, JSON.stringify( _action ) );
-		} catch ( e ) {
-			global.console.debug( 'DISPATCH', type, 'JSON ERROR' );
-		}
-
-		return next( action );
-	};
-}
-
-/**
- * Redux DevTools logger implementation.
- *
- * We set this because wp.data registry middleware is not extendable,
- * but it has built-in support for the devtools extension.
- * Since extensions can't be used in an E2E context, we're safe to hijack the global here.
- *
- * @see {@link https://github.com/WordPress/gutenberg/blob/2611a1df0a423dd22cbbabef8f2e87eb91b54bb2/packages/data/src/namespace-store/index.js#L124-L147}
- * @return {Object} Redux logger enhancer.
- */
-global.__REDUX_DEVTOOLS_EXTENSION__ = () => {
-	return applyMiddleware( createLoggerMiddleware );
-};
+// Bootstrap the logger as early as possible.
+setupReduxLogger();
